Add tests for rendezvous hook and database

+++ src/tribler/core/components/ipv8/rendezvous/db/orm_bindings/certificate.py
@@ -17,6 +17,6 @@ def define_binding(db):
         public_key = Required(bytes, index=True)
         total = Required(float)
         count = Required(int)
-        last_updated = Required(int)
+        last_updated = Required(float)
 
     return PeerScore
